SelectValidatorScreen as base component
