+++ src/module/staking/component/core/AddStakeModal/AddStakeModal.tsx
@@ -2,8 +2,7 @@ import { createBackdrop, ExposedBackdropProps } from "@peersyst/react-native-com
 import SetAmountStakeScreen from "module/staking/screen/SetAmountStakeScreen/SetAmountStakeScreen";
 import { useTranslate } from "module/common/hook/useTranslate";
 import StakeModal, { ModalTabs } from "../StakeModal/StakeModal";
-import SelectValidatorScreen from "module/staking/screen/SelectValidatorScreen/SelectValidatorScreen";
-import useGetAllValidators from "module/staking/query/useGetAllValidators";
+import AddStakeValidatorScreen from "module/staking/screen/SelectValidatorScreen/AddStakeValidatorScreen/AddStakeValidatorScreen";
 
 export enum AddStakeScreens {
     SET_AMOUNT,
@@ -13,26 +12,16 @@ export enum AddStakeScreens {
 const AddStakeModal = createBackdrop((props: ExposedBackdropProps) => {
     const translate = useTranslate();
 
-    const { data: validators, isLoading } = useGetAllValidators();
-
     const tabs: ModalTabs[] = [
         {
-            tabContent: <SetAmountStakeScreen />,
             title: translate("stake_your_near"),
             tabIndex: AddStakeScreens.SET_AMOUNT,
+            tabContent: <SetAmountStakeScreen />,
         },
         {
-            tabContent: (
-                <SelectValidatorScreen
-                    validators={validators}
-                    loading={isLoading}
-                    message={translate("select_new_validator")}
-                    onFinish={() => undefined}
-                    withSearch
-                />
-            ),
             title: translate("select_validator"),
             tabIndex: AddStakeScreens.SELECT_VALIDATOR,
+            tabContent: <AddStakeValidatorScreen />,
         },
         {
             tabContent: <></>,
