+++ src/module/staking/component/core/UnstakeModal/UnstakeModal.tsx
@@ -1,9 +1,8 @@
-import { createModal, ExposedBackdropProps, useSetTab } from "@peersyst/react-native-components";
+import { createModal, ExposedBackdropProps } from "@peersyst/react-native-components";
 import StakeModal, { ModalTabs } from "module/staking/component/core/StakeModal/StakeModal";
 import { useTranslate } from "module/common/hook/useTranslate";
-import SelectValidatorScreen from "module/staking/screen/SelectValidatorScreen/SelectValidatorScreen";
-import useGetStakingValidators from "module/staking/hook/useGetStakingValidators";
 import SetAmountStakeScreen from "module/staking/screen/SetAmountStakeScreen/SetAmountStakeScreen";
+import UnstakeValidatorScreen from "module/staking/screen/SelectValidatorScreen/UnstakeValidatorScreen/UnstakeValidatorScreen";
 
 export enum UnstakeModalScreens {
     SELECT_VALIDATOR,
@@ -12,22 +11,12 @@ export enum UnstakeModalScreens {
 
 const UnstakeModal = createModal(({ ...rest }: ExposedBackdropProps): JSX.Element => {
     const translate = useTranslate();
-    const setTab = useSetTab();
-
-    const { stakingValidators } = useGetStakingValidators();
 
     const unstakeModalTabs: ModalTabs[] = [
         {
             title: translate("select_validator"),
             tabIndex: UnstakeModalScreens.SELECT_VALIDATOR,
-            tabContent: (
-                <SelectValidatorScreen
-                    message={translate("select_unstake_validator")}
-                    validators={stakingValidators}
-                    loading={false}
-                    onFinish={() => setTab(UnstakeModalScreens.SET_AMOUNT)}
-                />
-            ),
+            tabContent: <UnstakeValidatorScreen />,
         },
         {
             title: translate("unstake_your_near"),
